+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 1.1.0
+
+- rename `delete` to `remove`
+- add http methods get, post, put, delete, patch
+- Add local auth methods register, login, forgotPassword, resetPassword, sendEmailConfirmation, me
+
 ## 1.0.0
 
 - API rewrite